Switch to 8 bit pooling index

+++ driver/mloPoolingHost.hpp
@@ -90,7 +90,7 @@ bool mloPoolingForwardRunHostAndVerify(
 	const _T * top_ptr,
 	bool do_backward,
 	size_t * mask_ptr,
-	uint16_t * mask_gpu,
+	uint8_t * mask_gpu,
 	double allowedEps
 	)
 {
@@ -171,14 +171,14 @@ bool mloPoolingForwardRunHostAndVerify(
 					if (!found)
 					{
 						res_index = std::numeric_limits<size_t>::max();
-						res_index_gpu = std::numeric_limits<uint16_t>::max();
+						res_index_gpu = std::numeric_limits<uint8_t>::max();
 					}
 					if (pooling_method == MLO_POOLING_OP_MAX)
 					{
 						mask_ptr[b*top_batch_stride + o * top_channel_stride + j * top_stride + i] = res_index;
 						if (do_backward)
 						{
-							uint16_t mg = mask_gpu[b*top_batch_stride + o * top_channel_stride + j * top_stride + i];
+							uint8_t mg = mask_gpu[b*top_batch_stride + o * top_channel_stride + j * top_stride + i];
 							if (mg != res_index_gpu) {
 								std::cout << "Mask mistmatch, gpu " << mg << " cpu " << res_index_gpu << "(" << res_index << ")" << std::endl;
 								match = false;
